added the confirmation dialog box; user will asked once prior to clearing previous scores

+++ app/src/main/java/com/adeel/funtriviaquiz/ViewScoresActivity.java
@@ -2,7 +2,10 @@ package com.adeel.funtriviaquiz;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.app.AlertDialog;
+import android.app.Dialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.view.Gravity;
@@ -41,6 +44,15 @@ public class ViewScoresActivity extends AppCompatActivity {
 
         // Get the previous scores, if any
         getPreviousScores();
+
+        // To create dummy data
+//        editor = previousScores.edit();
+//        editor.putInt("1", 42);
+//        editor.putInt("2", 32);
+//        editor.putInt("3", 22);
+//        editor.putInt("4", 12);
+//        editor.putInt("5", 02);
+//        editor.commit();
     }
 
     public void getPreviousScores() {
@@ -65,18 +77,40 @@ public class ViewScoresActivity extends AppCompatActivity {
 
     public void clearPreviousScores(View view)
     {
-        // TODO: Show user a confirmation dialog box before clearing the previous scores
+        // Before allowing user to clear the previous scores, display a confirmation dialog box
 
-        // Show user Toast notification that scores have been cleared
-        Toast.makeText(getApplicationContext(), "Previous Scores have been cleared", Toast.LENGTH_LONG).show();
+        // First, register the OnClickListener. This will be attached to the AlertDialog.Builder obj
+        DialogInterface.OnClickListener dOnClickListener = new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialogInterface, int buttonClicked) {
 
-        // Clears the whole file
-        editor = previousScores.edit();
-        editor.clear();
-        editor.commit();
+                // User pressed 'Yes', thus clear all the previous scores
+                if (buttonClicked == DialogInterface.BUTTON_POSITIVE) {
+                    // Clears the whole file
+                    editor = previousScores.edit();
+                    editor.clear();
+                    editor.commit();
 
-        // Clear the arrayAdapter so that the getPreviousScores function doesn't append to it
-        arrayAdapter.clear();
-        getPreviousScores();
+                    // Clear the arrayAdapter so that the getPreviousScores function doesn't append to it
+                    arrayAdapter.clear();
+                    getPreviousScores();
+
+                    // Show user Toast notification that scores have been cleared
+                    Toast.makeText(getApplicationContext(), "Previous Scores have been cleared", Toast.LENGTH_SHORT).show();
+                }
+
+                // No need to do anything else since previous scores won't be cleared
+                // in any other case.
+            }
+        };
+
+        AlertDialog.Builder builder = new AlertDialog.Builder(ViewScoresActivity.this);
+
+        // The same OnClickListener object will be attached to both responses since it is handling
+        // the both cases.
+        builder.setMessage("Are you sure you want to clear your previous scores?")
+                .setPositiveButton("Yes", dOnClickListener)
+                .setNegativeButton("No", dOnClickListener)
+                .show();
     }
 }
